+++ packages/common/src/core/schemas/internal/events/EventGalleryCardUiSchema.ts
@@ -318,6 +318,12 @@ export async function buildUiSchema(
                   },
                 ],
               },
+              {
+                type: "Slot",
+                options: {
+                  name: "add-content-slot",
+                },
+              },
             ],
           },
           {
@@ -364,14 +370,24 @@ export async function buildUiSchema(
                   },
                 },
               },
+              // TODO: Re-add this once https://github.com/Esri/calcite-design-system/issues/10152 & https://github.com/Esri/calcite-design-system/issues/6059 are resolved
+              // {
+              //   label: `{{${i18nScope}.appearance.showAdditionalInfo.label:translate}}`,
+              //   scope: "/properties/showAdditionalInfo",
+              //   type: "Control",
+              //   options: {
+              //     control: "hub-field-input-switch",
+              //     layout: "inline-space-between",
+              //   },
+              // },
               {
-                label: `{{${i18nScope}.appearance.showAdditionalInfo.label:translate}}`,
-                scope: "/properties/showAdditionalInfo",
-                type: "Control",
+                label: `{{${i18nScope}.appearance.layout.label:translate}}`,
                 options: {
-                  control: "hub-field-input-switch",
-                  layout: "inline-space-between",
+                  control: "hub-field-input-select",
+                  enum: { i18nScope: `${i18nScope}.appearance.layout` },
                 },
+                scope: "/properties/layout",
+                type: "Control",
               },
             ],
           },
@@ -410,6 +426,7 @@ export async function buildUiSchema(
                     actions: [
                       {
                         label: `{{${i18nScope}.options.openIn.notice.link:translate}}`,
+                        ariaLabel: `{{${i18nScope}.options.openIn.notice.linkAriaLabel:translate}}`,
                         icon: "launch",
                         href: "https://www.w3.org/TR/WCAG20-TECHS/G200.html",
                         target: "_blank",